completed map

+++ src/1_arrowkey_turn.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   2_rotate_camera.c                                  :+:      :+:    :+:   */
+/*   1_arrowkey_turn.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: simon <simon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 23:30:28 by simon             #+#    #+#             */
-/*   Updated: 2024/09/04 00:48:21 by simon            ###   ########.fr       */
+/*   Updated: 2024/09/08 18:39:37 by simon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@ static void
 	rotate_camera(
 		t_camera *camera)
 {
-	const float	previous_x = camera->dir_x;
-	const float	prev_plane_x = camera->plane_x;
-	const float	rm[2][2] = {camera->rm[0], camera->rm[1] * camera->sign_rotate,
+	const double	previous_x = camera->dir_x;
+	const double	prev_plane_x = camera->plane_x;
+	const double	rm[2][2] = {camera->rm[0], camera->rm[1] * camera->sign_rotate,
 		-camera->rm[1] * camera->sign_rotate, camera->rm[0]};
 
 	if (camera->sign_rotate == 0)
